extracting header class
