+++ client/src/components/HomePage.js
@@ -8,7 +8,7 @@ import ArrowDropDownIcon from '@material-ui/icons/ArrowDropDown';
 import ArrowDropUpIcon from '@material-ui/icons/ArrowDropUp';
 import { Redirect } from 'react-router-dom';
 import LeaguesPreview from './LeaguesPreview.js';
-
+import Header from './Header.js';
 const backend_url = require('../utils/backendUrl.js').backend_url;
 const get = require('../utils/requests.js').getRequest;
 const post = require('../utils/requests.js').postRequest;
@@ -19,16 +19,15 @@ class HomePage extends React.Component{
     this.state = {
       profilePicture: null,
       username: null,
-      arrowUp: false,
-      redirectLogin: false,
       leagues: [],
-      leagueInfo: {}
+      leagueInfo: {},
+      render: false
     }
   }
   componentDidMount(){
     get(backend_url+'/users/profile-details').then(response => {
       this.setState({
-        username: response.data['username']
+        username: response.data.username
       });
     }).catch(error => {
       this.setState({
@@ -58,62 +57,20 @@ class HomePage extends React.Component{
         leagueInfo: info
       });
     }).catch(error => {console.error(error)});
-    console.log(this.state.leagueInfo);
+    this.setState({render: true});
   }
-  onExpand = () => {
-    if(this.state.arrowUp){
-      this.setState({arrowUp: false});
-    }
-    else{
-      this.setState({arrowUp: true});
-    }
-  }
-  renderRedirect = () => {
-    if(this.state.redirectLogin){
-      return <Redirect to="/" />
-    }
-  }
-  logout = () => {
-    get(backend_url+'/users/logout').then(response => {
-      this.setState({redirectLogin: true});
-    }).catch(error => {console.error(error)})
-  }
+
   render(){
-    let arrowIconStyle = {verticalAlign: 'middle', margin: 'auto'}
-    let dropdownMenuStyle = {color: '#7702fa'}
-    let menuListStyle = {backgroundColor: '#E5E8E8', borderRadius: '1em'}
     return(
       <div className="App">
         <div className="side-column">
         </div>
         <div className="home-container">
-          <div className="top-bar">
-            <img className="profile-pic" src={this.state.profilePicture} alt="profile-picture" />
-            <div style={{paddingLeft: '1em'}}>
-              <p style={{color: '#7702fa', marginBlockEnd: 'auto'}}>welcome,</p>
-              <h3 onClick={this.onExpand} className="header-text">@{this.state.username}
-                {this.state.arrowUp ?
-                  <ArrowDropUpIcon style={arrowIconStyle}/>
-                  : <ArrowDropDownIcon style={arrowIconStyle}/>
-                }
-              </h3>
-              {this.state.arrowUp ?
-                <MenuList style={menuListStyle}>
-                  <MenuItem style={dropdownMenuStyle}>edit profile</MenuItem>
-                  <MenuItem onClick={this.logout} style={dropdownMenuStyle}>logout</MenuItem>
-                </MenuList>
-                : null
-              }
-            </div>
-            <div className="logo-container">
-              <img className="logo" src={require("../images/smalllogo.png")} />
-            </div>
+          <Header profilePicture={this.state.profilePicture} username={this.state.username} />
+          <div className="leagues-container">
+            <LeaguesPreview leagues={this.state.leagues} leagueInfo={this.state.leagueInfo} />
+          </div>
         </div>
-        <div className="leagues-container">
-          <LeaguesPreview leagues={this.state.leagues} leagueInfo={this.state.leagueInfo} />
-        </div>
-      </div>
-        {this.renderRedirect()}
       </div>
     );
   }
