+++ client/src/stylesheets/App.css
@@ -91,6 +91,10 @@
   flex-wrap: wrap;
   padding-left: 1em;
   padding-right: 2em;
+  margin-block-start: 0em;
+  margin-block-end: 0em;
+  margin-inline-end: 0em;
+  margin-inline-start: 0em;
   width: 100%;
   height: 100%;
   align-content: space-evenly;
